endpoint para obtener vehiculo asociado a un repartidor

+++ src/main/java/com/progra3/proyecto/repository/VehiculoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.progra3.proyecto.entity.Vehiculo;
 
-public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
+public interface VehiculoRepository extends JpaRepository<Vehiculo, Long> {
 	
 	List<Vehiculo> findByTipo(String tipo);
