+++ src/main/java/com/progra3/proyecto/servicejpa/VehiculoServiceImpl.java
@@ -22,7 +22,7 @@ public class VehiculoServiceImpl implements IVehiculoService {
     }
 
     @Override
-    public Vehiculo getById(Integer id) {
+    public Vehiculo getById(Long id) {
         Optional<Vehiculo> vehiculo = vehiculoRepository.findById(id);
         return vehiculo.orElse(null); 
     }
@@ -33,12 +33,12 @@ public class VehiculoServiceImpl implements IVehiculoService {
 	}
 
 	@Override
-	public void delete(Integer id) {
+	public void delete(Long id) {
 		vehiculoRepository.deleteById(id);
 	}
 	
 	@Override
-	public boolean exists(Integer id) {
+	public boolean exists(Long id) {
 		return id == null ? false : vehiculoRepository.existsById(id);
 	}
